+++ packages/frontend/src/App.svelte
@@ -6,10 +6,12 @@ import Route from './lib/Route.svelte';
 import Build from './Build.svelte';
 import { onMount } from 'svelte';
 import { getRouterState } from './api/client';
-import Homepage from './Homepage.svelte';
 import { rpcBrowser } from '/@/api/client';
 import { Messages } from '/@shared/src/messages/Messages';
 import DiskImageDetails from './lib/disk-image/DiskImageDetails.svelte';
+import Navigation from './Navigation.svelte';
+import DiskImagesList from './lib/disk-image/DiskImagesList.svelte';
+import Dashboard from './lib/dashboard/Dashboard.svelte';
 
 router.mode.hash();
 
@@ -22,7 +24,7 @@ onMount(() => {
   isMounted = true;
 
   return rpcBrowser.subscribe(Messages.MSG_NAVIGATE_BUILD, (x: string) => {
-    router.goto(`/build/${x}`);
+    router.goto(`/disk-images/build/${x}`);
   });
 });
 </script>
@@ -30,16 +32,22 @@ onMount(() => {
 <Route path="/*" breadcrumb="Bootable Containers" isAppMounted={isMounted} let:meta>
   <main class="flex flex-col w-screen h-screen overflow-hidden bg-[var(--pd-content-bg)]">
     <div class="flex flex-row w-full h-full overflow-hidden">
-      <Route path="/" breadcrumb="Bootable Containers">
-        <Homepage />
+      <Navigation meta={meta} />
+
+      <Route path="/" breadcrumb="Dashboard">
+        <Dashboard />
       </Route>
-      <Route path="/build" breadcrumb="Build">
-        <Build />
+      <Route path="/disk-images/" breadcrumb="Disk Images">
+        <DiskImagesList />
       </Route>
-      <Route path="/details/:id/*" breadcrumb="Disk Image Details" let:meta>
+      <Route path="/disk-image/:id/*" breadcrumb="Disk Image Details" let:meta>
         <DiskImageDetails id={meta.params.id} />
       </Route>
-      <Route path="/build/:name/:tag" breadcrumb="Build" let:meta>
+
+      <Route path="/disk-images/build" breadcrumb="Build">
+        <Build />
+      </Route>
+      <Route path="/disk-images/build/:name/:tag" breadcrumb="Build" let:meta>
         <Build imageName={decodeURIComponent(meta.params.name)} imageTag={decodeURIComponent(meta.params.tag)} />
       </Route>
     </div>
