+++ packages/frontend/src/lib/disk-image/DiskImageEmptyScreen.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+import { EmptyScreen } from '@podman-desktop/ui-svelte';
+
+import DiskImageIcon from '../DiskImageIcon.svelte';
+</script>
+
+<EmptyScreen icon={DiskImageIcon} title="No disk images" message="Click Build to create a new disk image" />
