+++ packages/frontend/src/lib/navigation.ts
@@ -1,0 +1,28 @@
+/**********************************************************************
+ * Copyright (C) 2024 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ***********************************************************************/
+import { router } from 'tinro';
+import { bootcClient } from '../api/client';
+
+export async function goToDiskImages(): Promise<void> {
+  router.goto('/disk-images');
+}
+
+export async function gotoBuild(): Promise<void> {
+  bootcClient.telemetryLogUsage('nav-build');
+  router.goto('/disk-images/build');
+}
